+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New `/lab/project-aidah` page
 - Button "Terapkan" on bottom sheet in "/lab/murajaah-at-taisir" page for better UX
+- Move title of header "/lab/murajaah-at-taisir" page to bottom sheet via info button for mobile display
 
 ### Changed